Allow base64 encoding of fixedsizebinary arrays (#18950)
## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax. For example
`Closes #123` indicates that this PR will close issue #123.
-->

- Closes #18949 

## Rationale for this change

We need to be able to base64 encode fixedsizebinary types.

## Are these changes tested?

Yes

## Are there any user-facing changes?

No

---------

Co-authored-by: Tim Saucer <timsaucer@gmail.com>