feat: Add constant column extraction and rewriting for projections in ParquetOpener (#19136)
## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax. For example
`Closes #123` indicates that this PR will close issue #123.
-->

- Closes #19089

## Rationale for this change

<!--
Why are you proposing this change? If this is already explained clearly
in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand
your changes and offer better suggestions for fixes.
-->

Use file/group statistics to detect constant (including all-NULL)
columns so we can avoid reading/decoding useless data and simplify
predicates for earlier pruning.

## What changes are included in this PR?

- Detect constant columns during Parquet scan setup, rewrite those
columns to Literals, shrink the projection mask, and fold constants into
predicates.
- Rebuild file pruner after constant folding to allow early pruning when
possible.

<!--
There is no need to duplicate the description in the issue here but it
is sometimes worth providing a summary of the individual changes in this
PR.
-->

## Are these changes tested?
Yes

<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?
-->

## Are there any user-facing changes?
No

<!--
If there are user-facing changes then we may require documentation to be
updated before approving the PR.
-->

<!--
If there are any breaking changes to public APIs, please add the `api
change` label.
-->

Co-authored-by: Adrian Garcia Badaracco <1755071+adriangb@users.noreply.github.com>