refactor: move metrics module to `datafusion-common` crate (#19247)
## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax. For example
`Closes #123` indicates that this PR will close issue #123.
-->

## Rationale for this change

<!--
Why are you proposing this change? If this is already explained clearly
in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand
your changes and offer better suggestions for fixes.
-->
https://github.com/apache/datafusion/issues/18456 requires this
refactor. Now we have dependency physical-plan -> physical-expr ->
common, and metrics module lives in physical-plan. We want metrics
module to directly accessible from physical-expr crate, so they should
be moved to common crate.

For compile time concern, I've checked `cargo build --profile
release-nonlto --timings` benchmark, compiling `common` crate takes 5s
in 3min total, so it has no noticeable impact. They're all keep in this
project wide root `datafusion-common` dependency for simplicity.

## What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it
is sometimes worth providing a summary of the individual changes in this
PR.
-->
- Metrics module depends on memory util `get_record_batch_memory_size`,
so it's first moved to `common` crate
- Move metrics module to `common` crate

Above moves are all re-exported in the original place to keep the public
APIs backward-compatible

## Are these changes tested?

<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?
-->

## Are there any user-facing changes?

<!--
If there are user-facing changes then we may require documentation to be
updated before approving the PR.
-->

<!--
If there are any breaking changes to public APIs, please add the `api
change` label.
-->