Add runtime config options for `list_files_cache_limit` and `list_files_cache_ttl` (#19108)
## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax. For example
`Closes #123` indicates that this PR will close issue #123.
-->

- Closes #19056

## Rationale for this change

<!--
Why are you proposing this change? If this is already explained clearly
in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand
your changes and offer better suggestions for fixes.
-->

Make the list file cache memory limit and TTL configurable via runtime
config.

## What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it
is sometimes worth providing a summary of the individual changes in this
PR.
-->

* Add ability to `SET` and `RESET` `list_files_cache_limit` and
`list_files_cache_ttl`
* `list_files_cache_ttl` will expect the duration to look like either
`1m30s` or `30` (I'm wondering if it would be simpler for it to just
accept a single unit?)
* Add `update_cache_ttl()` to the `ListFilesCache` trait so we can
update it from `RuntimeEnvBuilder::build()`
* Add config entries

## Are these changes tested?

<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?
-->
Yes

## Are there any user-facing changes?

<!--
If there are user-facing changes then we may require documentation to be
updated before approving the PR.
-->

`update_cache_ttl()` has been added to the `ListFilesCache` trait

<!--
If there are any breaking changes to public APIs, please add the `api
change` label.
-->

---------

Co-authored-by: Andrew Lamb <andrew@nerdnetworks.org>