Minor: Make `ProjectionExpr::new` easier to use with constants (#19343)
## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax. For example
`Closes #123` indicates that this PR will close issue #123.
-->

- Closes #.

## Rationale for this change


While reviewing the code from @pepijnve in
https://github.com/apache/datafusion/pull/19287 I noticed it was a
little awkward to create `ProjectionExpr` from constants. Let's add a
nicer signature for that.

Instead of
```rust
                ProjectionExpr::new(lit("0"), "l".to_string()),
```

Do
```rust
                ProjectionExpr::new(lit("0"), "l"),
```




## What changes are included in this PR?

1. Add new method and simplicy some tests

## Are these changes tested?
By CI
## Are there any user-facing changes?

A slightly nicer API