Implement disk spilling for all grouping ordering modes in GroupedHashAggregateStream (#19287)
## Which issue does this PR close?

- Closes #19286.
- Related to #13123

## Rationale for this change

GroupedHashAggregateStream currently always reports that it can spill to
the memory tracking subsystem even though this is dependent on the
aggregation mode and the grouping order.
The optimistic logic in `group_aggregate_batch` does not correctly take
the spilling preconditions into account which can lead to excessive
memory use.

In order to to resolve this, this PR implements disk spilling for all
grouping modes.

## What changes are included in this PR?

- Correctly set `MemoryConsumer::can_spill` to reflect actual spilling
behaviour
- Ensure optimistic out-of-memory tolerance in `group_aggregate_batch`
is aligned with disk spilling or early emission logic
- Implement output order respecting disk spilling for partially and
fully sorted inputs.

## Are these changes tested?

Added additional test case to demonstrate problem.
Added test case to check that output order is respected after spilling.

## Are there any user-facing changes?

Yes, memory exhaustion may be reported much earlier in the query
pipeline than is currently the case. In my local tests with a per
consumer memory limit of 32MiB, grouped aggregation would consume 480MiB
in practice. This was then reported by ExternalSortExec which choked on
trying to reserve that much memory.