Add recursive protection on planner's `create_physical_expr` (#19299)
## Which issue does this PR close?

None, as this PR is rather self-describing.

## Rationale for this change

We recently encountered stack overflow issues in the physical planner's
`create_physical_expr` function, when dealing with deeply nested
expression trees.

In the rest of the codebase, it seems the `recursive::recursive`
attribute is used to prevent this.

## What changes are included in this PR?

This PR adds the `recursive::recursive` attribute on
`create_physical_expr` whenever the `recursive_protection` feature is
enabled.

## Are these changes tested?

Yes, but mostly for the sake of manually testing the change.

I don't think it should be merged, as it has to be `#[ignore]`d anyway,
as the `recursive_protection` is not enabled in tests (which means it
would systematically crash).

I've left it in this PR however, to make review / manual testing easier.

## Are there any user-facing changes?

None.

---------

Co-authored-by: Andrew Lamb <andrew@nerdnetworks.org>