fix: csv schema_infer_max_records set to 0 return null datatype (#19432)
## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax. For example
`Closes #123` indicates that this PR will close issue #123.
-->

- close https://github.com/apache/datafusion/issues/19417

## Rationale for this change

- see https://github.com/apache/datafusion/issues/19417
- related to https://github.com/apache/datafusion/pull/17796

## What changes are included in this PR?

when schema_infer_max_records set to 0 in csv, return datatype as string

## Are these changes tested?

add test case for schema_infer_max_records equal to 0

## Are there any user-facing changes?

<!--
If there are user-facing changes then we may require documentation to be
updated before approving the PR.
-->

<!--
If there are any breaking changes to public APIs, please add the `api
change` label.
-->