Implement nested recursive CTEs (#18956)
It gives a name (the table name) to each `WorkTable`. This way
`WorkTableExec` can recognize its own `WorkTable`

Note that it doesn't allow multiple occurrences of the same CTE name:
it's not possible to implement things like "join with itself" correctly
with only the work table.

## Which issue does this PR close?

- Closes #18955.

## Rationale for this change

Support nested recursive CTEs without co-recursion. This is useful to
e.g. implement SPARQL or other graph query languages.

## What changes are included in this PR?

## Are these changes tested?

Yes! There is a nested recursive query in the test file

## Are there any user-facing changes?

Nested recursive queries are now allowed instead of failing with a "not
implemented" error