fix: CteWorkTable: properly apply TableProvider::scan projection argument (#18993)
It was previously ignored

## Which issue does this PR close?

- Closes #18992.

## Rationale for this change

All `TableProvider` implementations must support the `projection`
argument of the `scan` method. This was not the case in `CteWorkTable`.

## What changes are included in this PR?

Minimal implementation of the projection support. The projection applied
before the plan node return results. It might be nice to push it further
inside of the recursion implementation to reduce memory consumption but
I preferred to keep the fix minimal.

## Are these changes tested?

I have not figured out yet a nice SQL query to trigger an error without
this change. Some existing queries in `cte.slt` have set projection
(i.e. not `None`) so the code is very likely working. There is also a
test on the projection itself in `WorkTableExec`