Upgrade hashbrown to 0.16 (#19554)
## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax. For example
`Closes #123` indicates that this PR will close issue #123.
-->

- Closes #13433

## Rationale for this change

Keep up to date with hashbrown and possible (performance) improvements,
remove the `rawtable` functionality.


## What changes are included in this PR?

Upgrading, fixing case of nondeterminism in `HashSet` usage.

## Are these changes tested?

Yes existiing tests.

## Are there any user-facing changes?

No.