bugfix: preserve schema metadata for record batch in FFI (#19293)
## Which issue does this PR close?

- Closes #19292.

## Rationale for this change

Metadata on a schema for a record batch is lost during FFI conversion.
This is not always obvious because our other integrations like
`ExecutionPlan` and `TableProvider` have their own ways to provide the
schema.

This is currently an issue because we have FFI table providers who say
they have a specific schema but the schema of the actual record batches
does not match. The metadata is dropped.

## What changes are included in this PR?

We already have the schema in the FFI object, so use it both when
converting to and from FFI.

## Are these changes tested?

Unit test added.

## Are there any user-facing changes?

No