Enables DefaultListFilesCache by default (#19366)
## Which issue does this PR close?

- Closes https://github.com/apache/datafusion/issues/18827
- Closes https://github.com/apache/datafusion/issues/9654

## Rationale for this change

Now that the `DefaultListFilesCache` can be configured by users it's
safe to enable it by default and fix the tests that caching broke!

## What changes are included in this PR?

 - Sets the DefaultListFilesCache to be enabled by default
- Adds additional object store access tests to show list caching
behavior
 - Adds variable setting/reading sqllogic test cases
- Updates tests to disable caching when they relied on COPY commands so
changes can be detected for each query
 - Updates docs to help users upgrade


## Are these changes tested?

Yes, additional test cases have been added to help show the behavior of
the caching

## Are there any user-facing changes?

Yes, this changes the default behavior of DataFusion, however this
information is already captured in the upgrade guide.

##
cc @alamb

---------

Co-authored-by: Andrew Lamb <andrew@nerdnetworks.org>