feat: implement partition_statistics for WindowAggExec (#18534)
## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax. For example
`Closes #123` indicates that this PR will close issue #123.
-->

- Ref https://github.com/apache/datafusion/issues/15873

## Rationale for this change

<!--
Why are you proposing this change? If this is already explained clearly
in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand
your changes and offer better suggestions for fixes.
-->

## What changes are included in this PR?

Implemented partition_statistics for WindowAggExec.

## Are these changes tested?

Check the unit tests.

## Are there any user-facing changes?

No
<!--
If there are any breaking changes to public APIs, please add the `api
change` label.
-->