perf: Improve performance of `md5` (#19568)
## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax. For example
`Closes #123` indicates that this PR will close issue #123.
-->

- Part of https://github.com/apache/datafusion/issues/19569

## Rationale for this change

<!--
Why are you proposing this change? If this is already explained clearly
in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand
your changes and offer better suggestions for fixes.
-->

I ran microbenchmarks comparing DataFusion with DuckDB for string
functions (see https://github.com/apache/datafusion-benchmarks/pull/26)
and noticed that DF was very slow for `md5`.

This PR improves performance:

| Benchmark                  | Before | After  | Speedup     |
|----------------------------|--------|--------|-------------|
| md5_array (1024 strings)   | 206 µs | 100 µs | 2.1x faster |
| md5_scalar (single string) | 337 ns | 221 ns | 1.5x faster |

## What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it
is sometimes worth providing a summary of the individual changes in this
PR.
-->

Avoid using `write!` with a format string and use a more efficient
approach

## Are these changes tested?

<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?
-->

## Are there any user-facing changes?

<!--
If there are user-facing changes then we may require documentation to be
updated before approving the PR.
-->

<!--
If there are any breaking changes to public APIs, please add the `api
change` label.
-->